Clean up things before adding next stuff ...
