+++ src/renderer/services/chain/fees.ts
@@ -14,7 +14,33 @@ import * as BNB from '../binance/service'
 import * as BTC from '../bitcoin/context'
 import { selectedPoolAsset$, selectedPoolChain$ } from '../midgard/common'
 import { crossChainStakeMemo$ } from './memo'
-import { FeeLD, StakeFeesLD } from './types'
+import { FeeLD, LoadFeesHandler, StakeFeesLD } from './types'
+
+export const reloadFees = () => {
+  BNB.reloadFees()
+  BTC.reloadFees()
+}
+
+const reloadFeesByChain = (chain: Chain) => {
+  switch (chain) {
+    case 'BNB':
+      return BNB.reloadFees
+    case 'BTC':
+      return BTC.reloadFees
+    case 'ETH':
+      // reload ETH balances - not available yet
+      return () => {}
+    case 'THOR':
+      // reload THOR fees - not available yet
+      return () => {}
+    default:
+      return () => {}
+  }
+}
+
+export const reloadFees$: Rx.Observable<O.Option<LoadFeesHandler>> = selectedPoolChain$.pipe(
+  RxOp.map(O.map(reloadFeesByChain))
+)
 
 const reloadStakeFeesByChain = (chain: Chain) => {
   switch (chain) {
