+++ src/renderer/views/swap/SwapView.tsx
@@ -55,6 +55,9 @@ export const SwapView: React.FC<Props> = (_): JSX.Element => {
         poolAddresses,
         RD.map(A.head),
         RD.chain(rdFromOption(() => Error('No pool address available in list'))),
+        // TODO (@Veado)
+        // Do a health check for pool address before sending tx
+        // Issue #497: https://github.com/thorchain/asgardex-electron/issues/497
         // eslint-disable-next-line array-callback-return
         RD.map((endpoint) => {
           if (endpoint.address) {
